refactor: Endpoint Component combine set endpoint and fetch schema functionality in one component
